initial commit
everything here comes from https://github.com/katzenpost/katzenpost at commit
6a56194ee2089907ea4620478a1e90269a08584d

This commit is dated much after its initial children, which are being rebased
onto this after having been initially developed on top of a commit which was
equivalent to this (but with some more files we decided to drop here) but with
a filtered history from the katzenpost repo. We decided to drop the filtered
history due to it having a lot of misleading commit messages. So, this commit
now serves as the (human-readable) link to the history of these files.